requirements.txt
python-dotenv
# Windows API 연동을 위해 표준 라이브러리 ctypes를 사용하므로 추가 패키지는 필요하지 않으나,
# 향후 확장성을 고려해 필요한 패키지가 생기면 여기에 추가합니다.
